add spring security

+++ src/main/java/com/example/demo/controller/UserController.java
@@ -1,5 +1,10 @@
 package com.example.demo.controller;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,6 +19,7 @@ import org.springframework.ui.Model;
 import java.sql.SQLDataException;
 import com.example.demo.repository.UserRepository;
 import com.example.demo.repository.WishlistRepository;
+import com.example.demo.config.CustomUserDetails;
 import com.example.demo.model.User;
 
 @Controller
@@ -28,13 +34,9 @@ public class UserController {
 	}
 
 	@RequestMapping(value = {"/users", "/"}, method = RequestMethod.GET)
-	public String index(Model model) {
-		model.addAttribute("users", userRepository.findAll());
-		return "user/index";
-	}
+	public String show(Model model, @AuthenticationPrincipal CustomUserDetails authUser) {
+		long id = authUser.getId();
 
-	@GetMapping("/users/{id}")
-	public String show(Model model, @PathVariable("id") long id) {
 		model.addAttribute("user", userRepository.findOne(id));
 		model.addAttribute("wishlists", wishlistRepository.findWhere("user_id", id));
 		return "user/show";
@@ -46,8 +48,10 @@ public class UserController {
 		return "user/new";
 	}
 
-	@GetMapping("/users/{id}/edit")
-	public String edit(Model model, @PathVariable("id") long id) {
+	@GetMapping("/users/edit")
+	public String edit(Model model, @AuthenticationPrincipal CustomUserDetails authUser) {
+		long id = authUser.getId();
+
 		model.addAttribute("user", userRepository.findOne(id));
 		return "user/edit";
 	}
@@ -55,8 +59,10 @@ public class UserController {
 	@PostMapping("/users")
 	public String create(Model model, User user, RedirectAttributes redirectAttributes) {
 		try {
+            user.encodePassword();
 			user = userRepository.insert(user);
-			return "redirect:users/" + user.getId();
+			
+			return "redirect:users";
 		} catch (Exception e) {
 			redirectAttributes.addAttribute("error", e.getMessage());
 			return "redirect:users/new";
@@ -65,23 +71,30 @@ public class UserController {
 
 	@PatchMapping("/users")
 	public String update(Model model, User user, RedirectAttributes redirectAttributes) {
+		if (CustomUserDetails.notAllowed(user.getId()))
+			return "redirect:/users";
+
 		try {
+			user.encodePassword();
 			userRepository.update(user);
-			return "redirect:users/" + user.getId();
+			return "redirect:users";
 		} catch (Exception e) {
 			redirectAttributes.addAttribute("error", e.getMessage());
-			return "redirect:users/"  + user.getId() + "/edit";
+			return "redirect:users/edit";
 		}		
 	}
 	
 	@DeleteMapping("/users")
 	public String delete(Model model, User user, RedirectAttributes redirectAttributes) {
+		if (CustomUserDetails.notAllowed(user.getId()))
+			return "redirect:/users";
+
 		try {
 			userRepository.delete("id", user.getId());
 		} catch (Exception e) {
 			redirectAttributes.addAttribute("error", e.getMessage());
 		}
 
-		return "redirect:users";
+		return "redirect:login";
 	}
 }
